creative_intelligence/requirements.txt
# Core ML
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.40.0
peft>=0.10.0
accelerate>=0.28.0
datasets>=2.18.0
sentencepiece>=0.2.0

# Vision
Pillow>=10.0.0
clip-by-openai>=1.0.0
opencv-python>=4.9.0

# Tabular
xgboost>=2.0.0
lightgbm>=4.3.0
scikit-learn>=1.4.0
shap>=0.45.0

# Time series
prophet>=1.1.5

# Data
pandas>=2.2.0
numpy>=1.26.0
scipy>=1.12.0

# NLP evaluation
rouge-score>=0.1.2
nltk>=3.8.0

# Demo
gradio>=4.25.0
plotly>=5.20.0
matplotlib>=3.8.0

# Utils
pyyaml>=6.0
tqdm>=4.66.0
einops>=0.7.0
